react: Split Settings into separate component
